refactor(zudo-doc-v2): move filterHeaderRightItems into v2 with flags arg (#1729)
Move `src/utils/header-right-items.ts` to
`packages/zudo-doc-v2/src/header/right-items.ts`. The helper used to
read `settings` directly from `@/config/settings`; the new signature
takes a `HeaderRightItemFlags` object so callers translate host state
into the v2 boundary explicitly:

```
filterHeaderRightItems(items, {
  designTokenPanel, colorTweakPanel, aiAssistant,
  colorMode, hasLocales, hasVersions, hasGithubUrl,
})
```

This matches the inversion direction of the `<Header>` props
refactor: the renderer + its prefilter are both host-agnostic, and
the host wrapper computes the inputs in one place.
